+++ src/angular-cesium-widgets/angular-cesium-widgets.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { IconDragService } from './services/icon-drag.service';
+import { DraggableToMapService } from './services/draggable-to-map.service';
 import { CommonModule } from '@angular/common';
 import { AngularCesiumModule } from '../angular-cesium/angular-cesium.module';
 import { DraggableToMapDirective } from './directives/draggable-to-map';
@@ -12,7 +12,7 @@ import { DraggableToMapDirective } from './directives/draggable-to-map';
   ],
   exports: [DraggableToMapDirective],
   declarations: [DraggableToMapDirective],
-  providers: [IconDragService],
+  providers: [DraggableToMapService],
 })
 export class AngularCesiumWidgetsModule {
 }